+++ src/ValidationPackage.Codecs/Json/AuthorJson.fs
@@ -1,0 +1,7 @@
+namespace ValidationPackage.Codecs
+
+[<RequireQualifiedAccess>]
+module AuthorJson =
+
+    let encoder = Json.Encoders.Author.encode
+    let decoder = Json.Decoders.Author.decoder