+++ src/ValidationPackage.Codecs/Json/OntologyAnnotationJson.fs
@@ -1,0 +1,7 @@
+namespace ValidationPackage.Codecs
+
+[<RequireQualifiedAccess>]
+module OntologyAnnotationJson =
+
+    let encoder = Json.Encoders.OntologyAnnotation.encode
+    let decoder = Json.Decoders.OntologyAnnotation.decoder