+++ src/ValidationPackage.Codecs/Yaml/AuthorYaml.fs
@@ -1,0 +1,7 @@
+namespace ValidationPackage.Codecs
+
+[<RequireQualifiedAccess>]
+module AuthorYaml =
+
+    let internal encoder = Yaml.Encoders.Author.encode
+    let decoder = Yaml.Decoders.Author.decoder