+++ src/ValidationPackage.Codecs/Yaml/Decoders/Cwl.fs
@@ -1,19 +1,15 @@
-namespace ValidationPackage.Codecs
+namespace ValidationPackage.Codecs.Yaml.Decoders
 
 open System
 open YAMLicious
 open YAMLicious.YAMLiciousTypes
 open ValidationPackage.Model
+open ValidationPackage.Codecs.Yaml
 
 [<RequireQualifiedAccess>]
-module CwlYaml =
+module internal Cwl =
 
-    let internal commandInputTypeEncoder (inputType: CommandInputType) =
-        inputType
-        |> CommandInputType.toCwlString
-        |> YamlValue.string
-
-    let commandInputTypeDecoder element =
+    let commandInputType element =
         let value =
             match element with
             | YAMLElement.Value _
@@ -30,20 +26,7 @@ module CwlYaml =
         | :? ArgumentException ->
             invalidArg "element" $"unsupported CWL command input type: {value}"
 
-    let internal commandInputBindingEncoder (binding: CommandInputBinding) =
-        [
-            if binding.Position <> 0 then
-                "position", YamlValue.int binding.Position
-
-            if binding.Prefix <> "" then
-                "prefix", YamlValue.string binding.Prefix
-
-            if not binding.Separate then
-                "separate", YamlValue.bool binding.Separate
-        ]
-        |> YamlValue.object
-
-    let commandInputBindingDecoder element =
+    let commandInputBinding element =
         match element with
         | YAMLElement.Object _ ->
             Decode.object (fun get ->
@@ -67,32 +50,15 @@ module CwlYaml =
                 "element"
                 "CWL command input inputBinding must be a mapping"
 
-    let internal commandInputParameterEncoder (parameter: CommandInputParameter) =
-        [
-                "id", YamlValue.string parameter.Id
-                "type", commandInputTypeEncoder parameter.Type
-
-                if parameter.Label <> "" then
-                    "label", YamlValue.string parameter.Label
-
-                if parameter.Doc <> "" then
-                    "doc", YamlValue.string parameter.Doc
-
-                "inputBinding", commandInputBindingEncoder parameter.InputBinding
-        ]
-        |> YamlValue.object
-
-    let commandInputParameterDecoder element =
+    let commandInputParameter element =
         match element with
         | YAMLElement.Object _ ->
             Decode.object (fun get ->
                 let id = get.Optional.Field "id" Decode.string
-                let inputType = get.Optional.Field "type" commandInputTypeDecoder
+                let inputType = get.Optional.Field "type" commandInputType
 
                 let inputBinding =
-                    get.Optional.Field
-                        "inputBinding"
-                        commandInputBindingDecoder
+                    get.Optional.Field "inputBinding" commandInputBinding
 
                 match id, inputType, inputBinding with
                 | Some id, Some inputType, Some inputBinding ->
@@ -132,38 +98,12 @@ module CwlYaml =
                 "element"
                 "each CWL command input parameter must be a mapping"
 
-    let validateParameters (parameters: CommandInputParameter array) =
-        parameters
-        |> Array.iteri (fun index parameter ->
-            if String.IsNullOrWhiteSpace(parameter.Id) then
-                invalidArg
-                    "parameters"
-                    $"CWL command input parameter at index {index} requires a non-empty id"
-        )
-
-        parameters
-        |> Array.countBy (fun parameter -> parameter.Id)
-        |> Array.tryFind (fun (_, count) -> count > 1)
-        |> Option.iter (fun (id, _) ->
-            invalidArg
-                "parameters"
-                $"CWL command input parameter id must be unique, but was duplicated: {id}"
-        )
-
-        parameters
-
-    let internal commandInputParametersEncoder parameters =
-        parameters
-        |> validateParameters
-        |> Array.map commandInputParameterEncoder
-        |> YamlValue.array
-
-    let commandInputParametersDecoder element =
+    let commandInputParameters element =
         match element with
         | YAMLElement.Sequence _
         | YAMLElement.Object [ YAMLElement.Sequence _ ] ->
             element
-            |> Decode.array commandInputParameterDecoder
-            |> validateParameters
+            |> Decode.array commandInputParameter
+            |> CwlValidation.validateParameters
         | _ ->
             invalidArg "element" "AVPR Inputs must use the CWL array form"