+++ src/ValidationPackage.Codecs/Yaml/OntologyAnnotationYaml.fs
@@ -1,0 +1,7 @@
+namespace ValidationPackage.Codecs
+
+[<RequireQualifiedAccess>]
+module OntologyAnnotationYaml =
+
+    let internal encoder = Yaml.Encoders.OntologyAnnotation.encode
+    let decoder = Yaml.Decoders.OntologyAnnotation.decoder